+++ ham/api.ham/pom.xml
@@ -25,6 +25,12 @@
         </dependency>
         <dependency>
             <groupId>org.kendar</groupId>
+            <artifactId>app</artifactId>
+            <version>4.0.1</version>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.kendar</groupId>
             <artifactId>plugin.socks5.server</artifactId>
             <version>4.0.1</version>
             <scope>test</scope>
