+++ ham/api.ham/src/test/java/org/kendar/ham/fixes/HandlingOf5xxResponsesIT.java
@@ -1,28 +1,27 @@
 package org.kendar.ham.fixes;
 
 import com.sun.net.httpserver.HttpServer;
-import org.apache.commons.io.IOUtils;
 import org.apache.http.client.methods.HttpGet;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 import org.kendar.ham.*;
-import org.kendar.utils.Sleeper;
 
-import java.nio.charset.StandardCharsets;
 import java.util.HashMap;
 import java.util.Map;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotEquals;
 
 public class HandlingOf5xxResponsesIT {
     private static HamBuilder hamBuilder = (HamBuilder)GlobalSettings.builder();
     private static String proxyId;
+    private static HttpServer server;
 
     @BeforeAll
     public static void beforeAll() throws HamTestException, HamException {
+        server = getHttpServer(9091);
         HamStarter.runHamJar(HandlingOf5xxResponsesIT.class);
         proxyId = hamBuilder
                 .proxies()
@@ -32,64 +31,63 @@ public class HandlingOf5xxResponsesIT {
     @AfterAll
     public static void afterAll() throws HamException {
         hamBuilder.proxies().removeProxy(proxyId);
+        server.stop(0);
 
     }
 
     @ParameterizedTest
-    @ValueSource(ints = {200,201})
-    public void doTest2xx(int code) throws HamTestException, HamException {
-        startWithCode(code,null,code);
-    }
-
-    @ParameterizedTest
-    @ValueSource(ints = {304})
-    public void doTest3xx(int code) throws HamTestException, HamException {
-        startWithCode(code,null,code);
-    }
-
-    @ParameterizedTest
-    @ValueSource(ints = {401,404})
-    public void doTest4xx(int code) throws HamTestException, HamException {
-        startWithCode(code,null,code);
-    }
-
-    @ParameterizedTest
-    @ValueSource(ints = {500,501})
-    public void doTest5xx(int code) throws HamTestException, HamException {
-        startWithCode(code,null,code);
+    @ValueSource(ints = {200,201,401,404,500,501,304})
+    public void doTestA(int code) throws HamTestException, HamException {
+        System.out.println("Proposed "+code);
+        var httpGet = new HttpGet("http://www.local.test/testError?code="+code);
+        var clientResponse = hamBuilder.execute(httpGet);
+        assertEquals(code, clientResponse.getStatusLine().getStatusCode());
     }
 
 
 
     @ParameterizedTest
     @ValueSource(ints = {301,302})
-    public void doTest3xxRedirect(int code) throws HamTestException, HamException {
+    public void doTestB(int code) throws HamTestException, HamException {
+        System.out.println("Proposed "+code);
         var hd = new HashMap<String,String>();
         hd.put("Location","https://www.facebook.com");
-        startWithCode(code,hd,200);
+        var httpGet = new HttpGet("http://www.local.test/testError?code="+code+
+                "&location=http://www.local.test");
+        var clientResponse = hamBuilder.execute(httpGet);
+        assertNotEquals(code, clientResponse.getStatusLine().getStatusCode());
     }
 
-    private void startWithCode(int code, Map<String,String> headers,int expectedCode) throws HamTestException, HamException {
-        var httpServer =  getHttpServer(9091, code,headers);
-        try {
-            System.out.println("Started server with "+ code);
-            var httpGet = new HttpGet("http://www.local.test/testError");
-            var clientResponse = hamBuilder.execute(httpGet);
-            System.out.println("Retrieved response");
-            assertEquals(expectedCode, clientResponse.getStatusLine().getStatusCode());
-        }finally {
-            httpServer.stop(0);
-            Sleeper.sleep(200);
+    private static Map<String, String> queryToMap(String query) {
+        if(query == null) {
+            return null;
         }
+        Map<String, String> result = new HashMap<>();
+        for (String param : query.split("&")) {
+            String[] entry = param.split("=");
+            if (entry.length > 1) {
+                result.put(entry[0], entry[1]);
+            }else{
+                result.put(entry[0], "");
+            }
+        }
+        return result;
     }
 
-
-    private HttpServer getHttpServer(int gatewayPort,int throwEx,Map<String,String> headers) throws HamTestException {
+    private static HttpServer getHttpServer(int gatewayPort) throws HamTestException {
         var server = LocalHttpServer.startServer(gatewayPort,
                 new LocalHttpServer.LocalHandler("/testError", (call) -> {
                     try {
-                        System.out.println("Responding with "+throwEx);
-                        LocalHttpServer.sendResponse(call.httpExchange, throwEx, new byte[]{}, headers);
+
+                        System.out.println("Requested data");
+                        var headers = new HashMap<String,String>();
+                        var code = Integer.valueOf(call.query.get("code"));
+                        if(call.query.containsKey("location")){
+                            headers.put("Location",call.query.get("location"));
+                        }
+                        System.out.println("Responding with "+code);
+                        LocalHttpServer.sendResponse(call.httpExchange, code, new byte[]{}, headers);
+
                     } catch (Exception e) {
                         System.out.println(e.getMessage());
                     }
