+++ ham/api.js.builder/pom.xml
@@ -20,6 +20,12 @@
         </dependency>
         <dependency>
             <groupId>org.kendar</groupId>
+            <artifactId>app</artifactId>
+            <version>4.0.1</version>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.kendar</groupId>
             <artifactId>api.ham</artifactId>
             <version>4.0.1</version>
