+++ ham/api.test/src/test/java/org/kendar/ham/StarterIT.java
@@ -3,9 +3,9 @@ package org.kendar.ham;
 import org.junit.jupiter.api.Test;
 
 public class StarterIT {
-//    @Test
-//    public void doTest() throws HamTestException, InterruptedException {
-//        HamStarter.runHamJar(StarterIT.class);
-//        Thread.sleep(600000);
-//    }
+    @Test
+    public void doTest() throws HamTestException, InterruptedException {
+        HamStarter.runHamJar(StarterIT.class);
+        Thread.sleep(600000);
+    }
 }
